requirements.txt
fastapi==0.109.0
uvicorn==0.29.0
boto3==1.34.72
psycopg2-binary==2.9.9
sqlalchemy==2.0.29
